pyproject.toml
[project]
name = "billg-server"
version = "0.0.1"
description = "BillG Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.1",
    "fastapi[standard]>=0.115.12",
    "fastapi-utils>=0.8.0",
    "typing-inspect>=0.9.0",
    "openai>=1.75.0",
    "pyhumps>=3.8.0",
    "hvac>=2.3.0"
]

[project.optional-dependencies]
dev = [
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.1.0"
]